Some edits to clarify the message.
Typo check to be performed by another colleague. Minor edits to assure integrity of message, professional, friendly language accessible for everyone,  improvements to technical term translations, in a few limited instances, the way the message is provided in English requires indirect translation, the key is to maintain the core message and adopt to cultural context.